backend/requirements.txt
fastapi
uvicorn
httpx
python-multipart
pytest
pytest-asyncio
pytest-mock
pyyaml
motor
pymongo
python-dotenv
